Using rectangles to do the partitioning
Code now working nicely. Partitions are reasonable
when using rectangles. Some round off error has been
accounted for by adding a failsafe to expand the polygons
slightly if all nodes and elements (particularly elements)
cannot be partitioned. If this is the case, an updated
polygon file is written in the system directory so that
this process does not need to happen in the future.

Routines have been added to write the boundary conditions.

Removing executable attribute from files